Move profile utils to core
