+++ packages/app/src/resource/FormCreatePage.tsx
@@ -1,11 +1,12 @@
 import { Stack, Text } from '@mantine/core';
+import { addProfileToResource } from '@medplum/core';
 import { OperationOutcome, Resource } from '@medplum/fhirtypes';
 import { Document, ResourceForm, SupportedProfileStructureDefinition } from '@medplum/react';
 import { useCallback, useState } from 'react';
 import { useLocation, useParams } from 'react-router-dom';
-import { useCreateResource } from './useCreateResource';
 import { ProfileTabs } from './ProfileTabs';
-import { addProfileToResource, cleanResource } from './utils';
+import { useCreateResource } from './useCreateResource';
+import { cleanResource } from './utils';
 
 export function FormCreatePage(): JSX.Element {
   const { resourceType } = useParams();
